organize ticketstable templates
